Changed the action to take when create account button is clicked on signup activity such that it displays  a dialogbox instead of going to payment activity.

+++ app/src/main/java/com/anorlddroid/smartloan/database/UserEntity.kt
@@ -22,7 +22,7 @@ class UserEntity {
     var email: String? = null
 
     @ColumnInfo(name = "nationalID")
-    var nationalID: Int? = null
+    var nationalID: Long? = null
 
     @ColumnInfo(name = "password")
     var password: String? = null
